checking if classifier name exists now delegated to class Classifier
